capi: extract library finding logic to separate file
This reduce the bindings.py diff with the clang2py generated output,
making updating it more straightforward.
